% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmind.R
\name{addIPgeolocation}
\alias{addIPgeolocation}
\title{Given a vector of IP addresses it returns a data frame with
the IP addresses and its geolocation data (lat, long, accuracy and more).}
\usage{
addIPgeolocation(ips = "", df.maxmind = data.frame(), boost = FALSE)
}
\arguments{
\item{ips}{array of characters of IPv4 addresses}

\item{df.maxmind}{data frame from download.maxmind function}

\item{boost}{logical default set as FALSE, if TRUE it will use parallel computing using multiple cores}
}
\value{
data.frame
}
\description{
Given a vector of IP addresses it returns a data frame with
the IP addresses and its geolocation data (lat, long, accuracy and more).
}
\examples{
\dontrun{
geoips <- addIPgeolocation(ips = c("8.8.8.8", "147.81.23.1"),
                           df.maxmind = download.maxmind())
}
}
