% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scansio.R
\name{download.ftp.scans.io}
\alias{download.ftp.scans.io}
\title{Download and parse FTP connections from Rapid7 Open Data sonar project}
\usage{
download.ftp.scans.io(dirdata = "data")
}
\arguments{
\item{dirdata}{}
}
\value{
data.frame
}
\description{
Download and parse FTP connections from Rapid7 Open Data sonar project
}
\examples{
\dontrun{
raw.ftps <- download.ftp.scans.io()
}
}
