% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxmind.R
\name{download.maxmind}
\alias{download.maxmind}
\title{Download GeoLite2 data set that include geolocation information for network ranges (CIDRs)}
\usage{
download.maxmind(dirdata = "data")
}
\arguments{
\item{dirdata}{}
}
\value{
data.frame
}
\description{
Download GeoLite2 data set that include geolocation information for network ranges (CIDRs)
}
\examples{
\dontrun{
maxmind <- download.maxmind()
}
}
