% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scansio.R
\name{getScansioFTPs}
\alias{getScansioFTPs}
\title{Load FTP data set from Rapid7 and add IP geolocation data, providing a tidy data set.}
\usage{
getScansioFTPs(scope = 500, dirdata = "data", seed = 666)
}
\arguments{
\item{scope}{number of observations. Warning: if scope > 1000 it will use multiple cores.}

\item{dirdata}{path where raw data and tidy data frames will be stored}

\item{seed}{used to select the scope observations}
}
\description{
Load FTP data set from Rapid7 and add IP geolocation data, providing a tidy data set.
}
\examples{
\dontrun{
ftps <- getScansioFTPs()
}
}
